+++ lib/core/constants/strings.dart
@@ -1,5 +1,5 @@
 class Strings {
   // App
   static const appTitle = 'Flutter assignment';
-  static const appVersion = '1.1.1';
+  static const appVersion = '1.1.2';
 }