+++ lib/presentation/widgets/bike_tile.dart
@@ -9,7 +9,7 @@ import '../../core/constants/images.dart';
 import '../../logic/bloc/bike_bloc.dart';
 import '../utils/custom_text.dart';
 
-class BikeTile extends StatelessWidget {
+class BikeTile extends StatefulWidget {
   int index;
   BikeTile({
     super.key,
@@ -17,12 +17,17 @@ class BikeTile extends StatelessWidget {
   });
 
   @override
+  State<BikeTile> createState() => _BikeTileState();
+}
+
+class _BikeTileState extends State<BikeTile> {
+  @override
   Widget build(BuildContext context) {
     return BlocBuilder<AppThemeCubit, AppThemeState>(
       builder: (context, themeState) {
         return GestureDetector(
           onTap: (){
-            BlocProvider.of<BikeBloc>(context).add(BikeSelected(index: index, context: context));
+            BlocProvider.of<BikeBloc>(context).add(BikeSelected(index: widget.index, context: context));
           },
           child: SizedBox(
             child: Stack(
@@ -43,7 +48,7 @@ class BikeTile extends StatelessWidget {
                               Padding(
                                 padding: EdgeInsets.only(left: 3.w, bottom: 2.h, top: 7.h),
                                 child: CustomText(
-                                  textString: bikes[index]['Name'],
+                                  textString: bikes[widget.index]['Name'],
                                   fontFamily: 'faster one',
                                   textFontSize: 20.sp,
                                 ),
@@ -59,7 +64,7 @@ class BikeTile extends StatelessWidget {
                                   children: [
                                     Icon(Icons.local_gas_station, size: 20.sp,),
                                     CustomText(
-                                      textString: bikes[index]['Mileage'],
+                                      textString: bikes[widget.index]['Mileage'],
                                       textFontSize: 16.sp,
                                     ),
                                   ],
@@ -69,7 +74,7 @@ class BikeTile extends StatelessWidget {
                                   children: [
                                     Icon(Icons.account_balance_wallet, size: 20.sp,),
                                     CustomText(
-                                      textString: '₹${bikes[index]['Amount']}/D',
+                                      textString: '₹${bikes[widget.index]['Amount']}/D',
                                       textFontSize: 16.sp,
                                     ),
                                   ],
@@ -79,7 +84,7 @@ class BikeTile extends StatelessWidget {
                                   children: [
                                     Icon(Icons.stars, size: 20.sp,),
                                     CustomText(
-                                      textString: bikes[index]['Rating'],
+                                      textString: bikes[widget.index]['Rating'],
                                       textFontSize: 16.sp,
                                     ),
                                   ],
@@ -95,7 +100,7 @@ class BikeTile extends StatelessWidget {
                 Align(
                   alignment: Alignment.topCenter,
                   child: Image.asset(
-                    bikes[index]['URL'],
+                    bikes[widget.index]['URL'],
                     height: 20.h,
                   ),
                 ),